update watch function
