+++ controller/watcher/src/contracts/tx_hash.txt
@@ -1,0 +1,1 @@
+0x1d6a739b4516ff57d4beddd47e781816fcdd1f67822dafd1e95400cf05d3d0a1
