+++ controller/watcher/src/watch.py
@@ -8,37 +8,95 @@ from web3.middleware import geth_poa_middleware
 
 from control import servo_lock, servo_unlock
 
-with open('abi', 'r') as f:
-    abi = json.load(f)
+import tkinter
+from PIL import Image, ImageTk
 
-infura_url = 'http://geth:8545'
-tx_hash = "0xe5ae65b502bdea31300f3e1e5127428829a78b578cca343d914fd4c1a09d3a9a"
+time.sleep(10)
 
-w3 = Web3(Web3.HTTPProvider(infura_url))
-w3.middleware_onion.inject(geth_poa_middleware, layer=0)
 
-tx_receipt = w3.eth.getTransactionReceipt(tx_hash)
-contractAddress = tx_receipt['contractAddress']
-contract = w3.eth.contract(address=contractAddress, abi=abi)
-accounts = w3.eth.accounts
+def get_contract_address():
+    try:
+        with open('contracts/lock.json', 'r') as f:
+            abi = json.load(f)
+    except:
+        print('cannot import abi')
+
+    try:
+        with open('contracts/tx_hash.txt', 'r') as f:
+            tx_hash = f.read()
+    except:
+        ptint('cannot import tx_hash')
+
+    try:
+        infura_url = 'http://geth:8545'
+        w3 = Web3(Web3.HTTPProvider(infura_url))
+        w3.middleware_onion.inject(geth_poa_middleware, layer=0)
+        print(vars(w3.eth))
+    except:
+        print('cannot load web3 instance')
+
+    try:
+        tx_receipt = w3.eth.getTransactionReceipt(tx_hash)
+        contract_address = tx_receipt['contractAddress']
+        contract = w3.eth.contract(address=contract_address, abi=abi)
+        accounts = w3.eth.accounts
+    except Exception as e:
+        print('cannot load contract instance')
+        print(e)
+    
+    return contract_address
+
+
+def show_image():
+    global item, canvas
+
+    root = tkinter.Tk()
+    root.attributes('-fullscreen', True)
+    # root.bind('', lambda e: root.destroy())
+    root.title('Status')
+    root.geometry("1920x1080")
+    img = Image.open('image/display_locked_qr.jpeg')
+    img = ImageTk.PhotoImage(img)
+    canvas = tkinter.Canvas(bg="black", width=1920, height=1080)
+    canvas.place(x=0, y=0)
+    item = canvas.create_image(0, 0, image=img, anchor=tkinter.NW)
+    root.mainloop()
+
 
 def handle_event(event):
     receipt = w3.eth.waitForTransactionReceipt(event['transactionHash'])
     locked = contract.events.Locked().processReceipt(receipt)
     if locked:
+        img = Image.open('image/display_locked_qr.jpeg')
+        img = ImageTk.PhotoImage(img)
+        canvas.itemconfig(item, image=img)
         servo_lock()
         # print('locked')
+
     unlocked = contract.events.Unlocked().processReceipt(receipt)
     if unlocked:
+        img2 = Image.open('image/display_unlocked_qr.jpeg')
+        img2 = ImageTk.PhotoImage(img2)
+        canvas.itemconfig(item, image=img2)
         servo_unlock()
         # print('unlocked')
+    
 
-def log_loop(event_filter, poll_interval):
+def main():
+    thread1 = Thread(target=show_image)
+    thread1.start()
+    
+    interval = 2
     while True:
-        for event in event_filter.get_new_entries():
-            print(event)
-            handle_event(event)
-            time.sleep(poll_interval)
+        try:
+            contract_address = get_contract_address()
+            event_filter = w3.eth.filter({'fromBlock': 'latest', 'address': contract_address})
+            for event in event_filter.get_new_entries():
+                handle_event(event)
+        except:
+            print('cannot get event')
+        time.sleep(interval)
 
-block_filter = w3.eth.filter({'fromBlock':'latest', 'address':contractAddress})
-log_loop(block_filter, 2)
+
+if __name__ == '__main__':
+    main()
